requirements.txt
asgiref==3.3.1
Django==3.1.7
django-autocomplete-light==3.8.2
django-filter==2.4.0
django-polymorphic==3.0.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
Markdown==3.3.4
mysqlclient==2.0.3
Pillow==8.1.0
python-dotenv==0.15.0
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
